Configurando o Resource Server com a nova stack do Spring Security
