+++ src/main/java/com/course/springfood/core/security/ResourceServerConfig.java
@@ -1,0 +1,22 @@
+package com.course.springfood.core.security;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@EnableWebSecurity
+public class ResourceServerConfig extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .csrf().disable()  //Desativado CSRF
+                .authorizeRequests()
+                .anyRequest().authenticated()
+                .and()
+                .oauth2ResourceServer().opaqueToken();
+    }
+
+}
