+++ src/main/resources/application.properties
@@ -21,6 +21,11 @@ server.compression.enabled=true
 
 #logging.loggly.token=LOGGLY_CUSTOMER_TOKEN
 
+# Authorization Server
+spring.security.oauth2.resourceserver.opaquetoken.introspection-uri=http://localhost:8081/oauth/check_token
+spring.security.oauth2.resourceserver.opaquetoken.client-id=checktoken
+spring.security.oauth2.resourceserver.opaquetoken.client-secret=check123
+
 # Storage
 springfood.storage.tipo=local
 #springfood.storage.local.diretorio-fotos=DIRETORIO_DE_ARMAZENAMENTO_DAS_FOTOS
